requirements.txt
asgiref==3.8.1
attrs==25.1.0
black==24.10.0
class-registry==2.1.2
click==8.1.8
colorama==0.4.6
Django==5.1.4
django-debug-toolbar==5.0.1
django-extensions==3.2.3
django-filter==25.1
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.28.0
filters==1.3.2
flake8==7.1.2
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
rpds-py==0.23.0
six==1.17.0
sqlparse==0.5.3
tzdata==2024.2
uritemplate==4.1.1
